backend/requirements.txt
asgiref==3.9.1
Django==3.2.4
django-filter==23.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
